Update index.html
This is the version without the iframe embedding dribdat. 
I left in the iframe to other sites that I used for explanation, hope that is ok. 
I also left it up to the user to choose if they wanted to read about hackathons first or not because I think it is important to let each user determine their own sequence. I'm not sure if the Profile link works without me signed in; I gave a logged out link?
I hope the instructions won't be confusing if appended to dribdat. Some of them depend on location. If it doesn't work right, I have the old version which can be used separately.